requirements.txt
python-telegram-bot==21.0
google-generativeai==0.8.0
gspread==6.0.0
google-auth==2.27.0
python-dotenv==1.0.0
Pillow==10.2.0
google-api-python-client
matplotlib==3.8.2
requests==2.31.0
pandas==2.2.0
python-telegram-bot[job-queue]==21.0
openpyxl==3.1.2
